Harena

+++ public/assets/database/20260216-03-stockArgrnt.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS stockArgent (
     id INT AUTO_INCREMENT PRIMARY KEY,
     idVille INT NOT NULL UNIQUE,
-    quantite INT NOT NULL DEFAULT 0,
+    quantite BIGINT NOT NULL DEFAULT 0,
     dateCreation TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     dateModification TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     FOREIGN KEY (idVille) REFERENCES bngrc_villes(id) ON DELETE CASCADE
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS stockArgent (
 CREATE TABLE IF NOT EXISTS repartitionArgent (
     id INT AUTO_INCREMENT PRIMARY KEY,
     idBesoin INT NOT NULL,
-    montantReparti INT NOT NULL COMMENT 'Montant total réparti',
+    montantReparti BIGINT NOT NULL COMMENT 'Montant total réparti',
     dateRepartition TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (idBesoin) REFERENCES bngrc_besoins(id) ON DELETE CASCADE
 );
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS detailsRepartition (
     id INT AUTO_INCREMENT PRIMARY KEY,
     idRepartition INT NOT NULL,
     idVille INT NOT NULL,
-    montantRecu INT NOT NULL COMMENT 'Montant reçu lors de la répartition',
+    montantRecu BIGINT NOT NULL COMMENT 'Montant reçu lors de la répartition',
     FOREIGN KEY (idRepartition) REFERENCES repartitionArgent(id) ON DELETE CASCADE,
     FOREIGN KEY (idVille) REFERENCES bngrc_villes(id) ON DELETE CASCADE
 );
